Add tests, restructure
The final piece for 0.2.0 is adding tests to cover the documented usage,
which this does.

To make the testing easier I've split the actual meat of the code into
library file. This lets me test the `run` method directly. I could not
figure out how to test the code via `main`, and it seemed like doing so
wouldn't have been a good idea anyway.

During this move I pulled out Tokio. I think it's overkill for this. I
also thought it was causing testing problems, though it probably wasn't.
Testing problems or no, we don't need an async http core to this dumb
little app that downloads a single file from GitHub. I can do that with
just the `curl` crate.

Something I learned while writing tests, by default Rust parallellizes
its tests. Great! Except when you have three tests that are all trying
to write/read/delete the same file. So, dumb design on my part but I'm
fixing it for now by turning off test parallellization. If you run
`./script/test` everything is run in a single thread.

I've also added some other helper scripts and documented their usage in
the README.
