+++ script/test
@@ -1,0 +1,2 @@
+  script/update
+  RUST_TEST_THREADS=1 cargo test
