+++ script/update
@@ -1,0 +1,10 @@
+#!/bin/sh
+
+# script/update: Update application to run for its current checkout.
+
+set -e
+
+cd "$(dirname "$0")/.."
+
+script/setup
+cargo update
